#25652 coupon create+update
This will fix the issue where 2 webhooks are being run, coupon.created & coupon.updated when a coupon is created.

We're now checking and comparing the record created & record modified timestamp. A mismatch indicates an already existing coupon is being updated and hence we fire the coupon.updated hook at that instance.