#25652 - product & order create+update
This will fix the issue where 2 webhooks are being run, record.created & record.updated when  a record is created.

This commit fixes the issue for products & orders.

We're now checking and comparing the record created & record modified timestamp. A mismatch indicates an already existing record is being updated and hence we fire the record.updated hook at that instance.

Product:
Until the product is published, the created date is blanks, so the product.created webhook will be fired.

Once a product is published and then if the status is edited to draft and then republished in each instance the product.updated hook will be fired.